Amend the error thrown by the i18n library into a developer warning instead. Reasons:
The useFallback option actually means that the attempted load and fail is actually a valid procedure, and not one to throw an error.
The location of the error (in the generic get accessor for the locale storage) makes it pretty near impossible to catch this error.
The error thrown prevents subsequent ajax callbacks from firing (ajaxComplete, ajaxError, etc) and thus the fail cannot be picked up externally.